Adds WASM shim and benchmark integration
This adds a small `cdylib` shim library  around simdutf8 with cargo
configuration to target `wasm32-unknown-unknown`.

The benchmarks are augmented to optionally embed Wasmer and at compile
time build the WASM shim and embed it in the benchmarks, compile the
WASM module copy in the test slice and invoke the shim routine to
benchmark it.

A limitation of this current approach is that it measures the
overhead of calling across the WASM runtime boundary for each iteration.

Another limitation is that it is currently using the default cranelift
backend for Wasmer (which is the backend for wasmtime), but the LLVM
backend is more performant according to Wasmer's documentation.  The
benchmarks still allow you to get a reasonable set of expectation of
performance (relative numbers against std vs basic for example).