Cleanup as per PR feedback
Removes the `wasm32_simd128` feature in favor of controlling this
behavior with just `target_feature = simd128` which is consistent with
x86 `no_std`.

Updates the documentation to explain the requirement of the target
feature to drive selection and moves test/dev docs for WASM into its own
file as to not clutter the README.

Also makes one minor change to the bench build script to make clippy
happy.