Make sure there are no memcpy calls (#108)
Apparently even with the hand-rolled `memcpy_unaligned_nonoverlapping_inline_opt_lt_64` llvm still replaced that code for the last < 8 bytes with a call to `memcpy`. This fixes it.

The CI inlining tests now check for all present symbols. The rationale is that referenced symbols must be called from the simdutf8 if present.

As a drive-by change it uses `get_unchecked` to get rid of panicking code in `validate_utf8_at_offset` and `get_compat_error`.